created login system (first version)
